Update README.md
#Description:

This is the first stage in creating my professional portfolio.  I have created the basic framework for the page and put in placeholders for projects that are not yet completed.  By the end of the course, this website will be a professional portfolio that I can use to demonstrate my work.

#Screenshot:


#Link to Deployed Application: